[develop] Fix Jenkins Nightly Build (#1161)
The Jenkins nightly builds have been inconsistent or not working at all on the parallel works (PW) platforms. Some of the issues have been related to the instance's infrastructure on Azure or a conda conflict between the host machine and the conda built by the SRW App (originally seeing on GCP is now being observed for all PW platforms). This PR resolves the conda conflict by deactivating the host conda before activating the srw_app environment for all PW platforms. The solution for Azure requires configurations changes, which were done on the backend.

---------

Co-authored-by: Parallel Works app-run user <Edward.Snyder@mgmt-edwardsnyder-awsrocky8test-00006.pw-noaa-us-east-1.pw.local>
Co-authored-by: michael.lueken <Michael.Lueken@noaa.gov>