[develop] [bug fix]  Adding missing Intel variable for PW Azure  (#1167)
The SRW App still fails on the PW Azure instance. It appears that the compute node needs to be in the same zone as the controller node. To achieve this, the compute node instance type needs to change, which is failing because of a missing Intel variable. This update adds this missing Intel variable when running on Azure.

---------

Co-authored-by: michael.lueken <Michael.Lueken@noaa.gov>